+++ src/contribution-graph/ContributionGraph.tsx
@@ -1,34 +1,62 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { View } from "react-native";
-import { Svg, G, Text, Rect } from "react-native-svg";
 import _ from "lodash";
-import AbstractChart from "../abstract-chart";
+import React from "react";
+import { View } from "react-native";
+import { G, Rect, RectProps, Svg, Text } from "react-native-svg";
+
+import AbstractChart from "../AbstractChart";
+import { mapValue } from "../Utils";
+import {
+  convertToDate,
+  getBeginningTimeForDate,
+  shiftDate
+} from "./DateHelpers";
 import {
   DAYS_IN_WEEK,
   MILLISECONDS_IN_ONE_DAY,
   MONTH_LABELS
 } from "./constants";
-import {
-  shiftDate,
-  getBeginningTimeForDate,
-  convertToDate
-} from "./dateHelpers";
+import { ContributionGraphProps, ContributionGraphState } from ".";
 
 const SQUARE_SIZE = 20;
 const MONTH_LABEL_GUTTER_SIZE = 8;
 const paddingLeft = 32;
-class ContributionGraph extends AbstractChart {
-  constructor(props) {
+
+export type ContributionChartValue = {
+  value: number;
+  title: string;
+  tooltipDataAttrs: TooltipDataAttrs;
+  date: Date;
+};
+
+export type TooltipDataAttrs = (
+  value: ContributionChartValue
+) => Partial<RectProps> | Partial<RectProps>;
+
+class ContributionGraph extends AbstractChart<
+  ContributionGraphProps,
+  ContributionGraphState
+> {
+  constructor(props: ContributionGraphProps) {
     super(props);
+
+    let { maxValue, minValue, valueCache } = this.getValueCache(props.values);
+
     this.state = {
-      valueCache: this.getValueCache(props.values)
+      maxValue,
+      minValue,
+      valueCache
     };
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps: ContributionGraphProps) {
+    let { maxValue, minValue, valueCache } = this.getValueCache(
+      nextProps.values
+    );
+
     this.setState({
-      valueCache: this.getValueCache(nextProps.values)
+      maxValue,
+      minValue,
+      valueCache
     });
   }
 
@@ -92,67 +120,96 @@ class ContributionGraph extends AbstractChart {
     );
   }
 
-  getValueCache(values) {
-    return values.reduce((memo, value) => {
-      const date = convertToDate(value.date);
-      const index = Math.floor(
-        (date - this.getStartDateWithEmptyDays()) / MILLISECONDS_IN_ONE_DAY
-      );
-      memo[index] = {
-        value,
-        title: this.props.titleForValue
-          ? this.props.titleForValue(value)
-          : null,
-        tooltipDataAttrs: this.getTooltipDataAttrsForValue(value)
-      };
-      return memo;
-    }, {});
+  getValueCache(values: ContributionChartValue[]) {
+    let minValue = Infinity,
+      maxValue = -Infinity;
+
+    return {
+      valueCache: values.reduce((memo, value) => {
+        const date = convertToDate(value.date);
+
+        const index = Math.floor(
+          (date.valueOf() - this.getStartDateWithEmptyDays().valueOf()) /
+            MILLISECONDS_IN_ONE_DAY
+        );
+
+        minValue = Math.min(value[this.props.accessor], minValue);
+        maxValue = Math.max(value[this.props.accessor], maxValue);
+
+        memo[index] = {
+          value,
+          title: this.props.titleForValue
+            ? this.props.titleForValue(value)
+            : null,
+          tooltipDataAttrs: this.getTooltipDataAttrsForValue(value)
+        };
+
+        return memo;
+      }, {}),
+      minValue,
+      maxValue
+    };
   }
 
-  getValueForIndex(index) {
+  getValueForIndex(index: number) {
     if (this.state.valueCache[index]) {
       return this.state.valueCache[index].value;
     }
     return null;
   }
 
-  getClassNameForIndex(index) {
+  getClassNameForIndex(index: number) {
     if (this.state.valueCache[index]) {
       if (this.state.valueCache[index].value) {
-        const count = this.state.valueCache[index].value.count;
+        const count = this.state.valueCache[index].value[this.props.accessor];
+
         if (count) {
-          const opacity = (count * 0.15 > 1 ? 1 : count * 0.15) + 0.15;
+          const opacity = mapValue(
+            count,
+            this.state.minValue,
+            this.state.maxValue,
+            0.15 + 0.05, // + 0.05 to make smaller values a bit more visible
+            1
+          );
+
           return this.props.chartConfig.color(opacity);
         }
       }
     }
+
     return this.props.chartConfig.color(0.15);
   }
 
-  getTitleForIndex(index) {
+  getTitleForIndex(index: number) {
     if (this.state.valueCache[index]) {
       return this.state.valueCache[index].title;
     }
+
     return this.props.titleForValue ? this.props.titleForValue(null) : null;
   }
 
-  getTooltipDataAttrsForIndex(index) {
+  getTooltipDataAttrsForIndex(index: number) {
     if (this.state.valueCache[index]) {
       return this.state.valueCache[index].tooltipDataAttrs;
     }
-    return this.getTooltipDataAttrsForValue({ date: null, count: null });
+
+    return this.getTooltipDataAttrsForValue({
+      date: null,
+      [this.props.accessor]: null
+    } as ContributionChartValue);
   }
 
-  getTooltipDataAttrsForValue(value) {
+  getTooltipDataAttrsForValue(value: ContributionChartValue) {
     const { tooltipDataAttrs } = this.props;
 
     if (typeof tooltipDataAttrs === "function") {
       return tooltipDataAttrs(value);
     }
+
     return tooltipDataAttrs;
   }
 
-  getTransformForWeek(weekIndex) {
+  getTransformForWeek(weekIndex: number) {
     if (this.props.horizontal) {
       return [weekIndex * this.getSquareSizeWithGutter(), 50];
     }
@@ -180,14 +237,14 @@ class ContributionGraph extends AbstractChart {
     return `${this.getHeight()} ${this.getWidth()}`;
   }
 
-  getSquareCoordinates(dayIndex) {
+  getSquareCoordinates(dayIndex: number) {
     if (this.props.horizontal) {
       return [0, dayIndex * this.getSquareSizeWithGutter()];
     }
     return [dayIndex * this.getSquareSizeWithGutter(), 0];
   }
 
-  getMonthLabelCoordinates(weekIndex) {
+  getMonthLabelCoordinates(weekIndex: number) {
     if (this.props.horizontal) {
       return [
         weekIndex * this.getSquareSizeWithGutter(),
@@ -201,21 +258,18 @@ class ContributionGraph extends AbstractChart {
     ];
   }
 
-  handleClick(value) {
-    if (this.props.onClick) {
-      this.props.onClick(value);
-    }
-  }
-
-  renderSquare(dayIndex, index) {
+  renderSquare(dayIndex: number, index: number) {
     const indexOutOfRange =
       index < this.getNumEmptyDaysAtStart() ||
       index >= this.getNumEmptyDaysAtStart() + this.props.numDays;
+
     if (indexOutOfRange && !this.props.showOutOfRangeDays) {
       return null;
     }
+
     const [x, y] = this.getSquareCoordinates(dayIndex);
     const { squareSize = SQUARE_SIZE } = this.props;
+
     return (
       <Rect
         key={index}
@@ -225,12 +279,32 @@ class ContributionGraph extends AbstractChart {
         y={y}
         title={this.getTitleForIndex(index)}
         fill={this.getClassNameForIndex(index)}
+        onPress={() => {
+          this.handleDayPress(index);
+        }}
         {...this.getTooltipDataAttrsForIndex(index)}
       />
     );
   }
 
-  renderWeek(weekIndex) {
+  handleDayPress(index: number) {
+    if (!this.props.onDayPress) {
+      return;
+    }
+
+    this.props.onDayPress(
+      this.state.valueCache[index] && this.state.valueCache[index].value
+        ? this.state.valueCache[index].value
+        : {
+            [this.props.accessor]: 0,
+            date: new Date(
+              this.getStartDate().valueOf() + index * MILLISECONDS_IN_ONE_DAY
+            )
+          }
+    );
+  }
+
+  renderWeek(weekIndex: number) {
     const [x, y] = this.getTransformForWeek(weekIndex);
     return (
       <G key={weekIndex} x={x} y={y}>
@@ -251,13 +325,17 @@ class ContributionGraph extends AbstractChart {
     if (!this.props.showMonthLabels) {
       return null;
     }
+
     const weekRange = _.range(this.getWeekCount() - 1); // don't render for last week, because label will be cut off
+
     return weekRange.map(weekIndex => {
       const endOfWeek = shiftDate(
         this.getStartDateWithEmptyDays(),
         (weekIndex + 1) * DAYS_IN_WEEK
       );
+
       const [x, y] = this.getMonthLabelCoordinates(weekIndex);
+
       return endOfWeek.getDate() >= 1 && endOfWeek.getDate() <= DAYS_IN_WEEK ? (
         <Text
           key={weekIndex}
@@ -265,19 +343,37 @@ class ContributionGraph extends AbstractChart {
           y={y + 8}
           {...this.getPropsForLabels()}
         >
-          {MONTH_LABELS[endOfWeek.getMonth()]}
+          {this.props.getMonthLabel
+            ? this.props.getMonthLabel(endOfWeek.getMonth())
+            : MONTH_LABELS[endOfWeek.getMonth()]}
         </Text>
       ) : null;
     });
   }
 
+  public static defaultProps = {
+    numDays: 200,
+    endDate: new Date(),
+    gutterSize: 1,
+    squareSize: SQUARE_SIZE,
+    horizontal: true,
+    showMonthLabels: true,
+    showOutOfRangeDays: false,
+    accessor: "count",
+    classForValue: value => (value ? "black" : "#8cc665"),
+    style: {}
+  };
+
   render() {
-    const { style = {} } = this.props;
+    const { style } = this.props;
+
     let { borderRadius = 0 } = style;
+
     if (!borderRadius && this.props.chartConfig.style) {
       const stupidXo = this.props.chartConfig.style.borderRadius;
       borderRadius = stupidXo;
     }
+
     return (
       <View style={style}>
         <Svg height={this.props.height} width={this.props.width}>
@@ -300,44 +396,5 @@ class ContributionGraph extends AbstractChart {
     );
   }
 }
-
-ContributionGraph.ViewPropTypes = {
-  values: PropTypes.arrayOf(
-    // array of objects with date and arbitrary metadata
-    PropTypes.shape({
-      date: PropTypes.oneOfType([
-        PropTypes.string,
-        PropTypes.number,
-        PropTypes.instanceOf(Date)
-      ]).isRequired
-    }).isRequired
-  ).isRequired,
-  numDays: PropTypes.number, // number of days back from endDate to show
-  endDate: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-    PropTypes.instanceOf(Date)
-  ]), // end of date range
-  gutterSize: PropTypes.number, // size of space between squares
-  squareSize: PropTypes.number, // size of squares
-  horizontal: PropTypes.bool, // whether to orient horizontally or vertically
-  showMonthLabels: PropTypes.bool, // whether to show month labels
-  showOutOfRangeDays: PropTypes.bool, // whether to render squares for extra days in week after endDate, and before start date
-  tooltipDataAttrs: PropTypes.oneOfType([PropTypes.object, PropTypes.func]), // data attributes to add to square for setting 3rd party tooltips, e.g. { 'data-toggle': 'tooltip' } for bootstrap tooltips
-  titleForValue: PropTypes.func, // function which returns title text for value
-  classForValue: PropTypes.func, // function which returns html class for value
-  onClick: PropTypes.func // callback function when a square is clicked
-};
-
-ContributionGraph.defaultProps = {
-  numDays: 200,
-  endDate: new Date(),
-  gutterSize: 1,
-  squareSize: SQUARE_SIZE,
-  horizontal: true,
-  showMonthLabels: true,
-  showOutOfRangeDays: false,
-  classForValue: value => (value ? "black" : "#8cc665")
-};
 
 export default ContributionGraph;